+++ src/app/components/ExternalLink.tsx
@@ -1,6 +1,5 @@
-import { clsx } from 'clsx'
-
 import { assignInlineVars } from '@vanilla-extract/dynamic'
+import { clsx } from 'clsx'
 import { forwardRef } from 'react'
 import { useConfig } from '../hooks/useConfig.js'
 import * as styles from './ExternalLink.css.js'