+++ src/app/components/MobileSearch.tsx
@@ -1,12 +1,19 @@
 import * as Dialog from '@radix-ui/react-dialog'
 import { MagnifyingGlassIcon } from '@radix-ui/react-icons'
-import { useState } from 'react'
-
+import { useQueryState } from 'nuqs'
+import { useEffect, useState } from 'react'
+import { useMediaQuery } from '../hooks/useMediaQuery.js'
 import * as styles from './MobileSearch.css.js'
 import { SearchDialog } from './SearchDialog.js'
 
 export function MobileSearch() {
+  const [queryParam] = useQueryState('q', { defaultValue: '' })
   const [open, setOpen] = useState(false)
+
+  const matches = useMediaQuery('(max-width: 1080px)')
+  useEffect(() => {
+    if (matches && queryParam) setOpen(true)
+  }, [matches, queryParam])
 
   return (
     <Dialog.Root open={open} onOpenChange={setOpen}>
@@ -16,7 +23,7 @@ export function MobileSearch() {
         </button>
       </Dialog.Trigger>
 
-      <SearchDialog open={open} onClose={() => setOpen(false)} />
+      {matches && <SearchDialog open={open} onClose={() => setOpen(false)} />}
     </Dialog.Root>
   )
 }