+++ src/app/components/SearchDialog.tsx
@@ -10,13 +10,14 @@ import * as Label from '@radix-ui/react-label'
 import clsx from 'clsx'
 import { default as Mark } from 'mark.js'
 import type { SearchResult } from 'minisearch'
+import { useQueryState } from 'nuqs'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { Link, useNavigate } from 'react-router'
-
 import { useConfig } from '../hooks/useConfig.js'
 import { useDebounce } from '../hooks/useDebounce.js'
 import { useLocalStorage } from '../hooks/useLocalStorage.js'
 import { type Result, useSearchIndex } from '../hooks/useSearchIndex.js'
+import { useSessionStorage } from '../hooks/useSessionStorage.js'
 import { visuallyHidden } from '../styles/utils.css.js'
 import { Content } from './Content.js'
 import { KeyboardShortcut } from './KeyboardShortcut.js'
@@ -28,8 +29,20 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
   const inputRef = useRef<HTMLInputElement>(null)
   const listRef = useRef<HTMLUListElement>(null)
 
-  const [filterText, setFilterText] = useLocalStorage('filterText', '')
+  const [queryParam, setQueryParam] = useQueryState('q', { defaultValue: '' })
+  const [filterText, setFilterText] = useSessionStorage('filterText', queryParam)
   const searchTerm = useDebounce(filterText, 200)
+
+  useEffect(() => {
+    // Keep filter text in sync with query param.
+    if (queryParam) setFilterText(queryParam)
+  }, [queryParam, setFilterText])
+
+  useEffect(() => {
+    // Keep query param in sync with filter text.
+    setQueryParam(props.open ? (filterText ?? '') : '')
+  }, [filterText, setQueryParam, props.open])
+
   const searchIndex = useSearchIndex()
 
   const [selectedIndex, setSelectedIndex] = useState(-1)
@@ -124,7 +137,7 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
     return () => {
       window.removeEventListener('keydown', keyDownHandler)
     }
-  }, [navigate, resultsCount, setFilterText, selectedResult, props.open, props.onClose])
+  }, [navigate, resultsCount, selectedResult, props.open, props.onClose, setFilterText, props])
 
   useEffect(() => {
     if (searchTerm === '') return
@@ -170,6 +183,7 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
               width={20}
             />
           </Label.Root>
+          {/** biome-ignore lint/correctness/useUniqueElementIds: _ */}
           <input
             ref={inputRef}
             tabIndex={0}
@@ -215,11 +229,8 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
           )}
 
           {results.map((result, index) => (
-            // biome-ignore lint/a11y/useFocusableInteractive:
+            // biome-ignore lint/a11y/useAriaPropsSupportedByRole: _
             <li
-              // biome-ignore lint/a11y/noNoninteractiveElementToInteractiveRole:
-              // biome-ignore lint/a11y/useSemanticElements:
-              role="option"
               key={result.id}
               className={clsx(styles.result, index === selectedIndex && styles.resultSelected)}
               aria-selected={index === selectedIndex}
@@ -248,7 +259,7 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
                     .map((title: string) => (
                       <span className={styles.title} key={title}>
                         <span
-                          // biome-ignore lint/security/noDangerouslySetInnerHtml:
+                          // biome-ignore lint/security/noDangerouslySetInnerHtml: _
                           dangerouslySetInnerHTML={{ __html: title }}
                         />
                         <ChevronRightIcon className={styles.titleIcon} />
@@ -256,7 +267,7 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
                     ))}
                   <span className={styles.title}>
                     <span
-                      // biome-ignore lint/security/noDangerouslySetInnerHtml:
+                      // biome-ignore lint/security/noDangerouslySetInnerHtml: _
                       dangerouslySetInnerHTML={{ __html: result.title }}
                     />
                   </span>
@@ -266,7 +277,7 @@ export function SearchDialog(props: { open: boolean; onClose(): void }) {
                   <div className={styles.excerpt}>
                     <Content className={styles.content}>
                       <div
-                        // biome-ignore lint/security/noDangerouslySetInnerHtml:
+                        // biome-ignore lint/security/noDangerouslySetInnerHtml: _
                         dangerouslySetInnerHTML={{ __html: result.html }}
                       />
                     </Content>