+++ src/app/layouts/DocsLayout.tsx
@@ -1,8 +1,7 @@
+import { assignInlineVars } from '@vanilla-extract/dynamic'
 import clsx from 'clsx'
 import type { ReactNode } from 'react'
 import { useInView } from 'react-intersection-observer'
-
-import { assignInlineVars } from '@vanilla-extract/dynamic'
 import { bannerHeight } from '../components/Banner.css.js'
 import { Banner } from '../components/Banner.js'
 import { Content } from '../components/Content.js'
@@ -19,11 +18,7 @@ import { usePageData } from '../hooks/usePageData.js'
 import { contentVars, defaultFontFamily, fontFamilyVars } from '../styles/vars.css.js'
 import * as styles from './DocsLayout.css.js'
 
-export function DocsLayout({
-  children,
-}: {
-  children: ReactNode
-}) {
+export function DocsLayout({ children }: { children: ReactNode }) {
   const { banner, font } = useConfig()
   const { frontmatter = {} } = usePageData()
   const { content } = frontmatter