+++ src/vite/plugins/virtual-routes.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from 'node:fs'
+import { glob } from 'node:fs/promises'
 import { extname, resolve } from 'node:path'
-import { globby } from 'globby'
 import type { PluginOption } from 'vite'
 
 import { getGitTimestamp } from '../utils/getGitTimestamp.js'
@@ -10,7 +10,7 @@ export function virtualRoutes(): PluginOption {
   const virtualModuleId = 'virtual:routes'
   const resolvedVirtualModuleId = `\0${virtualModuleId}`
 
-  let glob: string
+  let globPattern: string
   let paths: string[] = []
 
   return {
@@ -38,7 +38,7 @@ export function virtualRoutes(): PluginOption {
         let code = 'export const routes = ['
         for (const path of paths) {
           const type = extname(path).match(/(mdx|md)/) ? 'mdx' : 'jsx'
-          const replacer = glob.split('*')[0]
+          const replacer = globPattern.split('*')[0]
 
           const content = readFileSync(path, 'utf-8')
           const filePath = path.replace(`${pagesPath}/`, '')
@@ -64,8 +64,8 @@ export function virtualRoutes(): PluginOption {
       const { config } = await resolveVocsConfig()
       const { rootDir } = config
       const pagesPath = resolve(rootDir, 'pages')
-      glob = `${pagesPath}/**/*.{md,mdx,ts,tsx,js,jsx}`
-      paths = await globby(glob)
+      globPattern = `${pagesPath}/**/*.{md,mdx,ts,tsx,js,jsx}`
+      paths = await Array.fromAsync(glob(globPattern))
     },
     handleHotUpdate() {
       // TODO: handle changes